refactor: move responder to own directory
