+++ src/components/IncidentDetails/AddResponders/ActionButtonGroup.tsx
@@ -1,0 +1,65 @@
+import clsx from "clsx";
+import { AddResponderAction } from "./AddResponder";
+
+type ActionButtonGroupProps = {
+  previousAction?: AddResponderAction;
+  nextAction?: AddResponderAction;
+} & React.HTMLProps<HTMLDivElement>;
+
+export function ActionButtonGroup({
+  previousAction,
+  nextAction,
+  className,
+  ...rest
+}: ActionButtonGroupProps) {
+  return (
+    <div
+      className={clsx(
+        "flex rounded-t-lg justify-between bg-gray-100 px-8 pb-4 items-end",
+        className
+      )}
+      {...rest}
+    >
+      <div className="flex flex-1">
+        {previousAction && (
+          <button
+            disabled={previousAction.disabled}
+            type="submit"
+            className={clsx(
+              !previousAction.primary
+                ? "inline-flex items-center px-4 py-2 border border-gray-300 shadow-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 sm:text-sm"
+                : "btn-primary",
+              "mt-4",
+              {
+                "btn-disabled": previousAction.disabled
+              }
+            )}
+            onClick={previousAction.handler}
+          >
+            {previousAction.label}
+          </button>
+        )}
+      </div>
+      <div className="flex flex-1 justify-end">
+        {nextAction && (
+          <button
+            disabled={nextAction.disabled}
+            type="submit"
+            className={clsx(
+              !nextAction.primary
+                ? "inline-flex items-center px-4 py-2 border border-gray-300 shadow-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 sm:text-sm"
+                : "btn-primary",
+              "mt-4",
+              {
+                "btn-disabled": nextAction.disabled
+              }
+            )}
+            onClick={nextAction.handler}
+          >
+            {nextAction.label}
+          </button>
+        )}
+      </div>
+    </div>
+  );
+}
