+++ src/components/IncidentDetails/ResponderTypes/AwsServiceRequest/AwsServiceRequest.stories.tsx
@@ -1,5 +1,5 @@
 import { useForm } from "react-hook-form";
-import { AddResponderFormValues } from "../../AddResponder";
+import { AddResponderFormValues } from "../../AddResponders/AddResponder";
 import { AwsServiceRequest } from "./index";
 
 export default {
