+++ src/components/IncidentDetails/ResponderTypes/AwsSupport/AwsSupport.stories.tsx
@@ -1,5 +1,5 @@
 import { useForm } from "react-hook-form";
-import { AddResponderFormValues } from "../../AddResponder";
+import { AddResponderFormValues } from "../../AddResponders/AddResponder";
 import { AwsSupport } from "./index";
 
 export default {
