+++ src/components/IncidentDetails/ResponderTypes/CA/CA.stories.tsx
@@ -1,5 +1,5 @@
 import { useForm } from "react-hook-form";
-import { AddResponderFormValues } from "../../AddResponder";
+import { AddResponderFormValues } from "../../AddResponders/AddResponder";
 import { CA } from "./index";
 
 export default {
