+++ src/components/IncidentDetails/ResponderTypes/Oracle/Oracle.stories.tsx
@@ -1,5 +1,5 @@
 import { useForm } from "react-hook-form";
-import { AddResponderFormValues } from "../../AddResponder";
+import { AddResponderFormValues } from "../../AddResponders/AddResponder";
 import { Oracle } from "./index";
 
 export default {
