+++ src/components/IncidentDetails/ResponderTypes/Oracle/index.tsx
@@ -6,7 +6,7 @@ import {
   UseFormSetValue
 } from "react-hook-form";
 import { TextInput } from "../../../TextInput";
-import { AddResponderFormValues } from "../../AddResponder";
+import { AddResponderFormValues } from "../../AddResponders/AddResponder";
 
 type OracleProps = {
   control: Control;
