+++ src/components/IncidentDetails/ResponderTypes/Person/Person.stories.tsx
@@ -1,5 +1,5 @@
 import { useForm } from "react-hook-form";
-import { AddResponderFormValues } from "../../AddResponder";
+import { AddResponderFormValues } from "../../AddResponders/AddResponder";
 import { Person } from "./index";
 
 export default {
