+++ src/components/IncidentDetails/ResponderTypes/Redhat/Redhat.stories.tsx
@@ -1,5 +1,5 @@
 import { useForm } from "react-hook-form";
-import { AddResponderFormValues } from "../../AddResponder";
+import { AddResponderFormValues } from "../../AddResponders/AddResponder";
 import { Redhat } from "./index";
 
 export default {
