+++ src/components/IncidentDetails/ResponderTypes/Redhat/index.tsx
@@ -6,7 +6,7 @@ import {
   UseFormSetValue
 } from "react-hook-form";
 import { TextInput } from "../../../TextInput";
-import { AddResponderFormValues } from "../../AddResponder";
+import { AddResponderFormValues } from "../../AddResponders/AddResponder";
 
 type RedhatProps = {
   control: Control;
