+++ src/components/IncidentDetails/ResponderTypes/ServiceNow/ServiceNow.stories.tsx
@@ -1,5 +1,5 @@
 import { useForm } from "react-hook-form";
-import { AddResponderFormValues } from "../../AddResponder";
+import { AddResponderFormValues } from "../../AddResponders/AddResponder";
 import { ServiceNow } from "./index";
 
 export default {
