+++ src/components/IncidentDetails/ResponderTypes/VMWare/VWWare.stories.tsx
@@ -1,5 +1,5 @@
 import { useForm } from "react-hook-form";
-import { AddResponderFormValues } from "../../AddResponder";
+import { AddResponderFormValues } from "../../AddResponders/AddResponder";
 import { VMWare } from "./index";
 
 export default {
